refactored select into it's own component
